requirements.txt
pymongo==3.4.0
requests==2.13.0
pyyaml==3.12
jieba
fasttext
numpy
flask
sklearn
scipy

